requirements.txt
rapidocr-onnxruntime==1.3.8
rapidocr-api==0.0.5
streamlit
pyclipper>=1.2.1
onnxruntime>=1.7.0
opencv_python>=4.5.1.48
numpy>=1.19.3
six>=1.15.0
Shapely>=1.7.1
PyYAML
Pillow